requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
spatialite==0.0.3
sqlparse==0.4.2
whitenoise==6.0.0
